prep entertainer details page for search
